requirements.txt
absl-py==0.2.0
astor==0.6.2
bleach==1.5.0
certifi==2018.1.18
chardet==3.0.4
coverage==4.5.1
cssselect==1.0.3
cssutils==1.0.2
Django==2.2.8
django-cors-headers==2.1.0
django-filter==2.2.0
django-premailer==0.2.0
django-simple-email-confirmation==0.60
djangorestframework==3.10.2
et-xmlfile==1.0.1
gast==0.2.0
grpcio==1.11.0
html5lib==0.9999999
idna==2.6
jdcal==1.3
lxml==4.2.5
Markdown==2.6.11
numpy==1.14.2
openpyxl==2.5.1
pandas==0.22.0
pip-autoremove==0.9.1
premailer==3.0.0
protobuf==3.5.2.post1
psycopg2==2.7.3.1
python-dateutil==2.7.2
pytz==2017.3
requests==2.18.4
six==1.11.0
sqlparse==0.3.0
tensorboard==1.7.0
tensorflow==1.7.0
termcolor==1.1.0
urllib3==1.22
Werkzeug==0.14.1
